*View model class added for all view. *View holder class added to all list item. *Glide v4 added. *Glide image loader helper class added.

+++ app/src/main/java/com/android/androiddatabinding/model/KnownFor.java
@@ -9,7 +9,7 @@ public class KnownFor {
 
     @SerializedName("vote_average")
     @Expose
-    private Integer voteAverage;
+    private float voteAverage;
     @SerializedName("vote_count")
     @Expose
     private Integer voteCount;
@@ -53,7 +53,7 @@ public class KnownFor {
     @Expose
     private String releaseDate;
 
-    public Integer getVoteAverage() {
+    public float getVoteAverage() {
         return voteAverage;
     }
